Rename Authenticated to QueryEditorWrapper
